+++ cpp/src/command_classes/DoorLockLogging.cpp
@@ -219,7 +219,7 @@ bool DoorLockLogging::RequestState
 bool DoorLockLogging::RequestValue
 (
 	uint32 const _requestFlags,
-	uint8 const _what,
+	uint16 const _what,
 	uint8 const _instance,
 	Driver::MsgQueue const _queue
 )
