+++ cpp/src/command_classes/Meter.cpp
@@ -180,7 +180,7 @@ bool Meter::RequestState
 bool Meter::RequestValue
 (
 	uint32 const _requestFlags,
-	uint8 const _dummy1,	// = 0 (not used)
+	uint16 const _dummy1,	// = 0 (not used)
 	uint8 const _instance,
 	Driver::MsgQueue const _queue
 )
