+++ cpp/src/command_classes/NodeNaming.cpp
@@ -234,7 +234,7 @@ bool NodeNaming::RequestState
 bool NodeNaming::RequestValue
 (
 	uint32 const _requestFlags,
-	uint8 const _getTypeEnum,
+	uint16 const _getTypeEnum,
 	uint8 const _instance,
 	Driver::MsgQueue const _queue
 )
