+++ cpp/src/command_classes/Security.cpp
@@ -161,7 +161,7 @@ bool Security::RequestState
 bool Security::RequestValue
 (
 		uint32 const _requestFlags,
-		uint8 const _index,
+		uint16 const _index,
 		uint8 const _instance,
 		Driver::MsgQueue const _queue
 )
