+++ cpp/src/command_classes/Security.h
@@ -85,7 +85,7 @@ namespace OpenZWave
 	private:
 		Security( uint32 const _homeId, uint8 const _nodeId );
 		bool RequestState( uint32 const _requestFlags, uint8 const _instance, Driver::MsgQueue const _queue);
-		bool RequestValue( uint32 const _requestFlags, uint8 const _index, uint8 const _instance, Driver::MsgQueue const _queue);
+		bool RequestValue( uint32 const _requestFlags, uint16 const _index, uint8 const _instance, Driver::MsgQueue const _queue);
 		bool HandleSupportedReport(uint8 const* _data, uint32 const _length);
 
 		bool m_schemeagreed;
