+++ cpp/src/command_classes/WakeUp.cpp
@@ -150,7 +150,7 @@ bool WakeUp::RequestState
 bool WakeUp::RequestValue
 (
 		uint32 const _requestFlags,
-		uint8 const _getTypeEnum,
+		uint16 const _getTypeEnum,
 		uint8 const _instance,
 		Driver::MsgQueue const _queue
 )
