+++ cpp/src/command_classes/ZWavePlusInfo.cpp
@@ -94,7 +94,7 @@ bool ZWavePlusInfo::RequestState
 bool ZWavePlusInfo::RequestValue
 (
 	uint32 const _requestFlags,
-	uint8 const _dummy1,	// = 0 (not used)
+	uint16 const _dummy1,	// = 0 (not used)
 	uint8 const _instance,
 	Driver::MsgQueue const _queue
 )
