+++ cpp/src/value_classes/ValueButton.cpp
@@ -46,7 +46,7 @@ ValueButton::ValueButton
 	ValueID::ValueGenre const _genre,
 	uint8 const _commandClassId,
 	uint8 const _instance,
-	uint8 const _index,
+	uint16 const _index,
 	string const& _label,
 	uint8 const _pollIntensity
 ):
