+++ cpp/src/value_classes/ValueDecimal.cpp
@@ -46,7 +46,7 @@ ValueDecimal::ValueDecimal
 	ValueID::ValueGenre const _genre,
 	uint8 const _commandClassId,
 	uint8 const _instance,
-	uint8 const _index,
+	uint16 const _index,
 	string const& _label,
 	string const& _units,
 	bool const _readOnly,
